Implement Matrix inverse operation (#6)
* Refactor for MultiDimensionalKalmanFilter
* Rename SystemState to SystemCycle
* Make EstimationFilter, KalmanFilter, and MeasureUpdatePredictFilter (and all their derived types) generic so OneDimensionalKalmanFilter can use double while MultiDimensionalKalmanFilter can use Vector
- update tests accordingly
* Begin Matrix inverse

* Implement left-side of inverse

* Implement augmented part of inverse

* Round to nearest thousandth

* Finish MatrixInverseOperation
* Use private methods so no need to implement more than
- testIsRunnable()
- testInverseWith5x5()
* MatrixInverseOperation is a Command object but does not implement Runnable because it throws an exception

+++ src/test/java/com/varunsingh/kalmanfilter/AlphaFilterTest.java
@@ -32,9 +32,9 @@ public class AlphaFilterTest {
     @Test
     public void testCalculateCurrentState() {
         // Arrange
-        int firstMeasurement = 1030;
-        int secondMeasurement = 989;
-        int thirdMeasurement = 1017;
+        double firstMeasurement = 1030;
+        double secondMeasurement = 989;
+        double thirdMeasurement = 1017;
 
         // Act
         alphaFilter.measure(firstMeasurement);
